refactor(bridge): extract presence diagnostics
